Move union-type action code into typed-actions module
