+++ src/typed-actions/utils.js
@@ -1,0 +1,13 @@
+"use strict";
+
+function isUnionTypeRecord(obj) {
+  return (
+    Array.isArray(obj._keys) &&
+    typeof obj._name === 'string' &&
+    typeof obj.case === 'function'
+  )
+}
+
+module.exports = {
+  isUnionTypeRecord,
+}
